@@ -15,13 +15,13 @@ publish_to: none
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   latlng: any
-  cached_network_image: any
+  cached_network_image: ^3.0.0
   map:
     path: ../
 
