@@ -28,15 +28,26 @@ class _MyHomePageState extends State<MyHomePage> {
     location: LatLng(35.68, 51.41),
   );
 
+  final markers = [
+    LatLng(35.674, 51.41),
+    LatLng(35.676, 51.41),
+    LatLng(35.678, 51.41),
+    LatLng(35.68, 51.41),
+    LatLng(35.682, 51.41),
+    LatLng(35.684, 51.41),
+    LatLng(35.686, 51.41),
+  ];
+
   void _gotoDefault() {
     controller.center = LatLng(35.68, 51.41);
   }
 
   void _onDoubleTap() {
     controller.zoom += 0.5;
+    setState(() {});
   }
 
-  Offset _dragStart;
+  Offset? _dragStart;
   double _scaleStart = 1.0;
   void _onScaleStart(ScaleStartDetails details) {
     _dragStart = details.focalPoint;
@@ -49,64 +60,96 @@ class _MyHomePageState extends State<MyHomePage> {
 
     if (scaleDiff > 0) {
       controller.zoom += 0.02;
+      setState(() {});
     } else if (scaleDiff < 0) {
       controller.zoom -= 0.02;
+      setState(() {});
     } else {
       final now = details.focalPoint;
-      final diff = now - _dragStart;
+      final diff = now - _dragStart!;
       _dragStart = now;
       controller.drag(diff.dx, diff.dy);
+      setState(() {});
     }
+  }
+
+  Widget _buildMarkerWidget(Offset pos, Color color) {
+    return Positioned(
+      left: pos.dx - 16,
+      top: pos.dy - 16,
+      width: 24,
+      height: 24,
+      child: Icon(Icons.location_on, color: color),
+    );
   }
 
   @override
   Widget build(BuildContext context) {
-    //final devicePixelRatio = MediaQuery.of(context).devicePixelRatio;
-    //controller.tileSize = 256 / devicePixelRatio;
-
     return Scaffold(
       appBar: AppBar(
         title: Text('Map Demo'),
       ),
-      body: GestureDetector(
-        onDoubleTap: _onDoubleTap,
-        onScaleStart: _onScaleStart,
-        onScaleUpdate: _onScaleUpdate,
-        onScaleEnd: (details) {
-          print(
-              "Location: ${controller.center.latitude}, ${controller.center.longitude}");
-        },
-        child: Stack(
-          children: [
-            MapLayoutBuilder(
-                controller: controller,
-                builder: (context, transformer) {
-                  return GestureDetector(
-                    onTapUp: (details) {
-                      final location = transformer
-                          .fromXYCoordsToLatLng(details.localPosition);
+      body: MapLayoutBuilder(
+        controller: controller,
+        builder: (context, transformer) {
+          final markerPositions =
+              markers.map(transformer.fromLatLngToXYCoords).toList();
 
-                      print('${location.longitude}, ${location.latitude}');
-                    },
-                    child: Map(
-                      controller: controller,
-                      builder: (context, x, y, z) {
-                        final url =
-                            'https://www.google.com/maps/vt/pb=!1m4!1m3!1i$z!2i$x!3i$y!2m3!1e0!2sm!3i420120488!3m7!2sen!5e1105!12m4!1e68!2m2!1sset!2sRoadmap!4e0!5m1!1e0!23i4111425';
+          final markerWidgets = markerPositions.map(
+            (pos) => _buildMarkerWidget(pos, Colors.red),
+          );
 
-                        return CachedNetworkImage(
-                          imageUrl: url,
-                          fit: BoxFit.cover,
-                        );
-                      },
-                    ),
-                  );
-                }),
-            Center(
-              child: Icon(Icons.close, color: Colors.red),
+          final homeLocation =
+              transformer.fromLatLngToXYCoords(LatLng(35.68, 51.412));
+
+          final homeMarkerWidget =
+              _buildMarkerWidget(homeLocation, Colors.black);
+
+          final centerLocation = Offset(
+              transformer.constraints.biggest.width / 2,
+              transformer.constraints.biggest.height / 2);
+
+          final centerMarkerWidget =
+              _buildMarkerWidget(centerLocation, Colors.purple);
+
+          return GestureDetector(
+            onDoubleTap: _onDoubleTap,
+            onScaleStart: _onScaleStart,
+            onScaleUpdate: _onScaleUpdate,
+            onTapUp: (details) {
+              final location =
+                  transformer.fromXYCoordsToLatLng(details.localPosition);
+
+              final clicked = transformer.fromLatLngToXYCoords(location);
+
+              print('${location.longitude}, ${location.latitude}');
+              print('${clicked.dx}, ${clicked.dy}');
+              print('${details.localPosition.dx}, ${details.localPosition.dy}');
+            },
+            child: Stack(
+              children: [
+                Map(
+                  controller: controller,
+                  builder: (context, x, y, z) {
+                    final url =
+                        'https://www.google.com/maps/vt/pb=!1m4!1m3!1i$z!2i$x!3i$y!2m3!1e0!2sm!3i420120488!3m7!2sen!5e1105!12m4!1e68!2m2!1sset!2sRoadmap!4e0!5m1!1e0!23i4111425';
+
+                    return CachedNetworkImage(
+                      imageUrl: url,
+                      fit: BoxFit.cover,
+                    );
+                  },
+                ),
+                Positioned(
+                  child: Icon(Icons.close, color: Colors.black),
+                ),
+                homeMarkerWidget,
+                ...markerWidgets,
+                centerMarkerWidget,
+              ],
             ),
-          ],
-        ),
+          );
+        },
       ),
       floatingActionButton: FloatingActionButton(
         onPressed: _gotoDefault,
